changed double precision values to single precision values (does not do such a difference for my computer but probably will for others)

+++ activation.h
@@ -2,31 +2,31 @@
 
 #include <stddef.h>
 
-typedef void (*Activation_func)(double*); 
+typedef void (*Activation_func)(float*); 
 
-typedef void (*Output_Activation_func)(size_t , double*);
+typedef void (*Output_Activation_func)(size_t , float*);
 
-typedef double (*Output_Activation_derivative)(double);
+typedef float (*Output_Activation_derivative)(float);
 
-typedef double (*Activation_derivative)(double);
+typedef float (*Activation_derivative)(float);
 
-void relu(double* x);
+void relu(float* x);
 
-void leaky_relu(double* x);
+void leaky_relu(float* x);
 
-double d_relu(double x);
+float d_relu(float x);
 
-double d_leaky_relu(double x);
+float d_leaky_relu(float x);
 
-void sigmoid(double* x);
+void sigmoid(float* x);
 
-double d_sigmoid(double x);
+float d_sigmoid(float x);
 
-void softmax(size_t nb_values , double* values);
+void softmax(size_t nb_values , float* values);
 
-double d_softmax(double x);
+float d_softmax(float x);
 
-void sigmoid_output(size_t nb_values, double* values);
+void sigmoid_output(size_t nb_values, float* values);
 
 Activation_derivative resolve_derivative(Activation_func func);
