+++ eval.c
@@ -3,7 +3,7 @@
 #include "eval.h"
 
 
-double NN_error_rate(size_t nb_examples , int* predictions, int* labels){
+float NN_error_rate(size_t nb_examples , int* predictions, int* labels){
 
     size_t incorrect = 0;
 
@@ -16,11 +16,11 @@ double NN_error_rate(size_t nb_examples , int* predictions, int* labels){
         
     }
     
-    return ((double) incorrect)/nb_examples; 
+    return ((float) incorrect)/nb_examples; 
 
 }
 
-double NN_accuracy_score(size_t nb_examples , int* predictions, int* labels){
+float NN_accuracy_score(size_t nb_examples , int* predictions, int* labels){
 
     return 1.0 - NN_error_rate(nb_examples, predictions, labels);
